feat: moved to react-grid-layout on player/[username]
